OptoGscope with main body notch
Modified the design for the main body so that the original baseplate design can be used. A notch is made into the side of the body to allow for the set screw extension to fit in snugly into the bottom of the body.
The different files that end in different decimal points indicate how much wiggle room there is between the set screw extension and the size of the notch in the main body- units are in mm.

No other files were modified. The assembly files used the new designs for the main body, but the old designs for the baseplate, focus slider, and cap and the v0 design for the filter cover.